Add files via upload
The code provided was written on Matlab and can be used to align photometry and behavioral data and to make calculations on calcium signal around behavioral events.
We used this code to analyze data recorded with the Neurophotometrics Photometry System and behavioral data analyzed with the Noldus Observer software. 
Sample data is provided to run the code.